chore(engine): Moved some network-only code to NetworkPlayer
